Rename typedb.rs to server.rs
